Renamed class handling assembling transactions for contracts to ContractsAdapter
