+++ ts/identityWallet/types.ts
@@ -1,12 +1,12 @@
 import { Identity } from '../identity/identity'
 import { IVaultedKeyProvider } from '../vaultedKeyProvider/types'
 import { IKeyMetadata } from '../credentials/signedCredential/types'
-import {IContracts, IContractsGateway} from '../contracts/types'
+import {IContractsAdapter, IContractsGateway} from '../contracts/types'
 
 export interface IIdentityWalletCreateArgs {
   vaultedKeyProvider: IVaultedKeyProvider
   identity: Identity
   publicKeyMetadata: IKeyMetadata
-  contractHandler: IContracts
+  contractHandler: IContractsAdapter
   contractConnector: IContractsGateway
 }
